Generalize report_encryptor.py into document_encryptor.py and extract the sensitive fields to report_encryptor.py
